scripts/requirements-docs.txt
# Python dependencies for documentation extraction pipeline
# Install with: pip install -r scripts/requirements-docs.txt

# YAML parsing for frontmatter
pyyaml>=6.0

# Markdown processing
markdown>=3.5

# File pattern matching
pathspec>=0.11.0
